fix Column.write
This fixes the writing of long values into large columns.
