+++ arithmetization/src/main/java/net/consensys/linea/zktracer/lt/Column.java
@@ -1,4 +1,20 @@
+/*
+ * Copyright ConsenSys Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
+ * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations under the License.
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
 package net.consensys.linea.zktracer.lt;
+
+import java.nio.ByteBuffer;
 
 import net.consensys.linea.zktracer.Trace;
 
@@ -60,9 +76,9 @@ public interface Column extends Trace.Column {
   }
 
   /**
-   * Provides a static encoding of column data, where each element is stored explicitly (i.e. not as
-   * a pool index). This uses an optimised encoding (e.g. where binary colunms are stored as bits)
-   * and, hence, is suitable only for the v2 file format.
+   * Provides an encoding for small column data, where each element is stored explicitly (i.e. not
+   * as a pool index). This uses an optimised encoding based upon the largest element encountered.
+   * This is suitable for the v2 file format.
    */
   class Small extends Base implements Column {
     private final long longMax;
@@ -110,6 +126,10 @@ public interface Column extends Trace.Column {
     }
   }
 
+  /**
+   * Provides an encoding for potentially large column data, where each element is stored as an
+   * index into the heap. This is suitable for the v2 file format.
+   */
   class Large extends Base implements Column {
     private final BytesHeap heap;
     private final int[] buffer;
@@ -128,25 +148,7 @@ public interface Column extends Trace.Column {
 
     @Override
     public void write(long value) {
-      byte b3 = (byte) (value >> 24);
-      byte b2 = (byte) (value >> 16);
-      byte b1 = (byte) (value >> 8);
-      byte b0 = (byte) value;
-      byte[] bytes;
-      // ensure bytes in trimmed and in big endian form.
-      if (value <= 0xffff) {
-        if (value <= 0xff) {
-          bytes = new byte[] {b0};
-        } else {
-          bytes = new byte[] {b1, b0};
-        }
-      } else if (value <= 0xffffff) {
-        bytes = new byte[] {b2, b1, b0};
-      } else {
-        bytes = new byte[] {b3, b2, b1, b0};
-      }
-      //
-      this.write(bytes);
+      this.write(Util.long2TruncatedBytes(value));
     }
 
     /**
@@ -166,6 +168,92 @@ public interface Column extends Trace.Column {
      */
     public Encoding toEncoding() {
       return Encoding.ofPool(buffer, bitwidth());
+    }
+  }
+
+  /**
+   * Provides a simple encoding of column data, where each element is stored directly in place. This
+   * is suitable for the v1 file format.
+   */
+  class Raw implements Trace.Column {
+    private final String name;
+    private final int bitWidth;
+    private final int byteWidth;
+    private final long longMax;
+    private final ByteBuffer buffer;
+
+    public Raw(String name, int bitwidth, int length) {
+      this.name = name;
+      this.bitWidth = bitwidth;
+      this.byteWidth = Util.byteWidth(bitwidth);
+      this.longMax = 1L << bitwidth;
+      this.buffer = ByteBuffer.allocate(length * byteWidth);
+    }
+
+    @Override
+    public void write(boolean value) {
+      this.buffer.put((byte) (value ? 1 : 0));
+    }
+
+    @Override
+    public void write(long value) {
+      // Sanity check
+      if (longMax <= value) {
+        throw new IllegalArgumentException(name + " has invalid value (" + value + ")");
+      }
+      //
+      switch (byteWidth) {
+        case 8:
+          this.buffer.put((byte) (value >> 56));
+        case 7:
+          this.buffer.put((byte) (value >> 48));
+        case 6:
+          this.buffer.put((byte) (value >> 40));
+        case 5:
+          this.buffer.put((byte) (value >> 32));
+        case 4:
+          this.buffer.put((byte) (value >> 24));
+        case 3:
+          this.buffer.put((byte) (value >> 16));
+        case 2:
+          this.buffer.put((byte) (value >> 8));
+        case 1:
+          this.buffer.put((byte) value);
+          break;
+        default:
+          throw new IllegalArgumentException(name + " has invalid width (" + byteWidth + "bytes)");
+      }
+    }
+
+    /**
+     * Write element bytes
+     *
+     * @param bytes stored in big-endian form and already trimmed.
+     */
+    @Override
+    public void write(byte[] bytes) {
+      final int n = Util.bitLengthOf(bytes);
+      // Sanity check
+      if (n > bitWidth || bytes.length > byteWidth) {
+        throw new IllegalArgumentException(name + " has invalid width (" + n + " bits)");
+      }
+      // Write padding (if necessary)
+      for (int i = bytes.length; i < byteWidth; i++) {
+        buffer.put((byte) 0);
+      }
+      // Write data
+      for (int i = 0; i != bytes.length; i++) {
+        buffer.put(bytes[i]);
+      }
+    }
+
+    /**
+     * Access the underling array of bytes for this column.
+     *
+     * @return
+     */
+    public byte[] toBytes() {
+      return buffer.array();
     }
   }
 }
