+++ arithmetization/src/main/java/net/consensys/linea/zktracer/lt/LtFile.java
@@ -12,7 +12,6 @@
  *
  * SPDX-License-Identifier: Apache-2.0
  */
-
 package net.consensys.linea.zktracer.lt;
 
 import java.io.IOException;
