+++ arithmetization/src/main/java/net/consensys/linea/zktracer/lt/Util.java
@@ -34,6 +34,53 @@ public class Util {
   }
 
   /**
+   * Convert a long value into an array of at most 8 bytes which are truncated (i.e. no leading
+   * zeros). Thus, the value 0 is returned as the empty array, etc.
+   *
+   * @param value
+   * @return
+   */
+  static byte[] long2TruncatedBytes(long value) {
+    final byte b7 = (byte) (value >> 56);
+    final byte b6 = (byte) (value >> 48);
+    final byte b5 = (byte) (value >> 40);
+    final byte b4 = (byte) (value >> 32);
+    final byte b3 = (byte) (value >> 24);
+    final byte b2 = (byte) (value >> 16);
+    final byte b1 = (byte) (value >> 8);
+    final byte b0 = (byte) value;
+    // Determine length
+    int len = 0;
+    while (value != 0) {
+      value = value >>> 8;
+      len++;
+    }
+    // Create array
+    byte[] bytes = new byte[len];
+    int index = 0;
+    switch (len) {
+      case 8:
+        bytes[index++] = b7;
+      case 7:
+        bytes[index++] = b6;
+      case 6:
+        bytes[index++] = b5;
+      case 5:
+        bytes[index++] = b4;
+      case 4:
+        bytes[index++] = b3;
+      case 3:
+        bytes[index++] = b2;
+      case 2:
+        bytes[index++] = b1;
+      case 1:
+        bytes[index++] = b0;
+    }
+    // Done
+    return bytes;
+  }
+
+  /**
    * Determine the minimal number of bits required to store the value held in a given set of bytes
    * (assuming a big endian layout). For example, a value of 0x0AFF (binary 0b00001010_11111111).
    * has a bit length of 12.
